PROJFBPRO-69: Refactor functions into update_mitigation_status.py
