+++ src/handle_data.py
@@ -10,7 +10,7 @@ from src.ea_rules_access_functions import get_hardening_variation_key_by_id
 from src.ea_rules_access_functions import get_hardening_id
 from src.ea_rules_access_functions import get_rule_ids
 from src.get_the_state_of_the_system import get_the_state
-from src.check_mapping import check_mapping
+from src.update_mitigation_status import check_mapping
 from src.update_mitigation_status import update_mitigation_status
